Basic data aggregation demonstration
To focus on the features that Spring Cloud provides we need to have an
explicit link between the customer app and the store app. This commit
exposes an additional endpoint that provide the profile of a customer.

A profile aggregates the data from a customer with its nearest stores.
The nearest stores are computed based on its current "location" using
a call on a search endpoint.

Right now the URL of the store app service is hard-coded which is
obviously a serious issue. We will fix that in further commits.