Eureka demonstration
Integrate Eureka (service registry) so that the customer app no longer
needs to know where the store app is located. Each app registers itself
to the registry (by default on `localhost:8761`).

Adding the starter also auto-configures a `RestTemplate` that is able
to translate application names to actual locations using the discovery
client. This effectively allows to replace the host by the name of the
application.

The location of Eureka and some eureka-specific config are added to the
config server (see branch `step2`)