Hystrix demonstration
What can we do if an app goes down? Should we cascade the failure or
return a degraded answer?

Hystrix allows you to decide that on an operation per operation basis. In
this demonstration, we provide a fallback method for the `StoreService`
so that no nearest stores are found if the store service is down.

This commit also adds a new module that exposes a management dashboard
with all kinds of statistics for a particular circuit. We can hammer the
`profiles` endpoint to show what happens when the store service is down.