CloudFoundry demonstration
This commit shows how the application can be pushed to the cloud using
CloudFoundry.

First, the config server is located at `http://localhost:8888` by default
so we need to provide an alternate location using an environment variable
that we can configure externally.

We then configure the cloudfoundry maven plugin to push any app with two
environment variables:

* `PROJECT_PREFIX` is the prefix to add to all apps to avoid name
conflicts. This is only for demo purposes
* `CONFIG_SERVER_URI` is the location of the config server

The configuration has to be updated as well to provide some "cloud" bits:

* Each client must provide its hostname and non secure port so that the
registry can properly provide an access to it
* The location of the Eureka server now needs to be provided as we were
using the default (`http://localhost:8767` until now)